Avoid race between zpool import and service config file generation
11.3-U2 added a sanity check when generating share configuration to
avoid adding shares to non-existent paths. In some cases configuration
file generation may happen before zpool import completes. This causes
shares on the pool with slow import to not be visible.

To work around this issue in 11.3 and keep change set minimal, check the
system ready state when generating service configuration files and short-
circuit the check for whether the share path exists if the system is
not ready.